* loader_wav.cpp: When a file fails to open, the system error 	message is now printed on stderr, instead of merely "Cannot open 	file".
	* sample_viewer.cpp: Fixed a bug where positions within a vector
	were calculated with floating-point numbers, resulting in
	out-of-bounds errors. Fixes bug #2120835.

	* various files: Added #include directives required by newer
	versions of GCC.

	


git-svn-id: https://svn.code.sf.net/p/cheesetracker/code@125 88f81a63-fe31-0410-a212-96c38ac4dcd5

+++ cheesetracker-qt3/common/interface__QT/audio/sample_viewer.cpp
@@ -368,8 +368,8 @@ int Sample_Viewer::get_factor() {
 }
 
 void Sample_Viewer::screen_to_sample(int p_int, float *p_max_peak,float *p_min_peak, int sample_cache_idx,int cur_width,int *p_final_sample) {
-	float sample_idx;
-	float size=sample_data->get_size();
+	size_t sample_idx;
+	size_t size=sample_data->get_size();
 	float factor=1;
 
 	if (sample_cache_idx!=-1) { //have cache
